+++ test/ham_fisted/hash_map_test.clj
@@ -99,55 +99,6 @@
         ))))
 
 
-(deftest split-iterator-test
-  (let [hm (HashMap.)
-        _ (dotimes [idx 2]
-            (.put hm idx idx))]
-    (is (= 2 (->> (map iterator-seq (.splitKeys hm 8))
-                  (map count)
-                  (reduce +)))))
-  (let [hm (HashMap.)
-        _ (dotimes [idx 10]
-            (.put hm idx idx))]
-    (is (= 10 (->> (map iterator-seq (.splitKeys hm 8))
-                  (map count)
-                  (reduce +)))))
-  ;; Force splitting in between nodes.
-  (let [hm (HashMap.)
-        _ (dotimes [idx 60]
-            (.put hm idx idx))]
-    (is (= 60 (->> (map iterator-seq (.splitKeys hm 7))
-                   (map count)
-                   (reduce +)))))
-  (let [hm (HashMap.)
-        _ (dotimes [idx 100]
-            (.put hm idx idx))]
-    (is (= 100 (->> (map iterator-seq (.splitKeys hm 8))
-                    (map count)
-                    (reduce +)))))
-  (let [hm (HashMap.)
-        _ (dotimes [idx 1000]
-            (.put hm idx idx))]
-    (is (= 1000 (->> (map iterator-seq (.splitKeys hm 8))
-                     (map count)
-                     (reduce +)))))
-  (let [hm (HashMap.)
-        _ (dotimes [idx 1000]
-            (.put hm idx idx))
-        _ (.put hm nil :a)]
-    (is (= 1001 (->> (map iterator-seq (.splitKeys hm 8))
-                     (map count)
-                     (reduce +)))))
-  (let [hm (HashMap.)
-        _ (dotimes [idx 1000]
-            (.put hm idx idx))
-        _ (.put hm nil :a)]
-    ;;Ensure we get split buckets at the root level and ensure we only
-    (is (= 1001 (->> (map iterator-seq (.splitKeys hm 13))
-                     (map count)
-                     (reduce +))))))
-
-
 
 (defmacro benchmark-us
   [op]